updating to double precision

+++ src/aalcalc_rust/src/collections.rs
@@ -62,13 +62,13 @@ use super::processes::{calculate_standard_deviation, calculate_st_deviation_two}
 #[derive(Debug)]
 pub struct SummaryStatistics {
     pub summary_id: i32,
-    pub ni_loss: f32,
-    pub ni_loss_squared: f32,
+    pub ni_loss: f64,
+    pub ni_loss_squared: f64,
     pub sample_size: i32,
-    pub total_loss: f32,
-    pub squared_total_loss: f32,
-    pub ni_loss_map: HashMap<i32, f32>,
-    pub period_categories: HashMap<i32, Vec<f32>>
+    pub total_loss: f64,
+    pub squared_total_loss: f64,
+    pub ni_loss_map: HashMap<i32, f64>,
+    pub period_categories: HashMap<i32, Vec<f64>>
 }
 
 impl SummaryStatistics {
@@ -78,8 +78,8 @@ impl SummaryStatistics {
     /// # Arguments
     /// * summary_id: the ID of the summaries that statistics are going to be collected on
     pub fn new(summary_id: i32) -> Self {
-        let ni_loss_map: HashMap<i32, f32> = HashMap::new();
-        let period_categories: HashMap<i32, Vec<f32>> = HashMap::new();
+        let ni_loss_map: HashMap<i32, f64> = HashMap::new();
+        let period_categories: HashMap<i32, Vec<f64>> = HashMap::new();
         return SummaryStatistics{
             summary_id,
             ni_loss: 0.0,
@@ -101,7 +101,7 @@ impl SummaryStatistics {
     /// # Arguments
     /// * number_of_periods: the ```period_number``` of the constructed ```OccurrenceData```
     pub fn print_type_one_stats(&self, number_of_periods: i32) {
-        let type_one_ni = self.ni_loss / number_of_periods as f32;
+        let type_one_ni = self.ni_loss / number_of_periods as f64;
         let standard_deviation = calculate_standard_deviation(&self.ni_loss_map, number_of_periods);
         println!("{},1,{},{}", self.summary_id, type_one_ni, standard_deviation);
     }
@@ -117,7 +117,7 @@ impl SummaryStatistics {
     /// * number_of_periods: the ```period_number``` of the constructed ```OccurrenceData```
     pub fn print_type_two_stats(&self, number_of_periods: i32) {
         let denominator  = (self.sample_size * number_of_periods);
-        let type_two_sample = self.total_loss / denominator as f32;
+        let type_two_sample = self.total_loss / denominator as f64;
         let standard_deviation_two = calculate_st_deviation_two(&self.period_categories, denominator);
         println!("{},2,{},{}", self.summary_id, type_two_sample, standard_deviation_two);
     }
